completed savings flex pending review

+++ lib/app/savings/modules/flex/model/data/flex_saving_config_request_body.dart
@@ -27,7 +27,7 @@ class FlexSavingConfigRequestBody {
   final FlexSaveMode? flexSaveMode;
   final String? contributionWeekDay;
   final int? contributionMonthDay;
-  final int? contributionAmount;
+  final double? contributionAmount;
   final int? customerAccountId;
   final int? customerId;
   final int? customerFlexSavingId;
